solve a bug in libblob run at 32 platform

+++ src/miscellany/buffer_convert.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
-#include <miscellany/addr_convert.h>
+#include <miscellany/buffer_convert.h>
 
 void addr_to_buffer(void *addr,uint8_t *buffer)
 {
 	unsigned long data = (unsigned long)addr;
 	int i;
 
-	for(i = 0; i < sizeof(int *); i++){
+	for(i = 0; i < (int)sizeof(int *); i++){
 		buffer[i] = data >> 8 * (sizeof(int *) - 1 - i);
 	}
 }
@@ -17,7 +17,7 @@ void *buffer_to_addr(uint8_t *buffer)
 	unsigned long d = 0, t = 0;
 	int i;
 
-	for ( i = 0; i < sizeof(int *); i++){
+	for ( i = 0; i < (int)sizeof(int *); i++){
 		t = buffer[i];
 		d |= t << 8 * (sizeof(int *) - 1 - i); 
 	}
